Use tab based indenting and pip for dependencies
Most of the indentation was using spaces, which works, but is a source
of subtle/hard to find bugs. The repository now also contains a
requirements.txt file for use with pip, so that we can avoid placing
dependencies under version control within our own code repositories.
You can install the required libraries for the project by running:

$: pip install -r requirements.txt

from the root directory (the one with the requirements.txt in it)