Add test data for order 1-8 handling
The original script was written in a way so that it can handle the
new 1-8 orders without change. stimuli_order7.xlsx is identical to
stimuli.xlsx except that its order is 7 rather than 3. Running that
file through the script with the new datasource template (with all 8
orders) produces an output with the correctly corresponding orders and
carriers pulled out.